Will move the use cases to bank package
