--- bank/transactions_show/transactions_test.go
@@ -1,20 +1,0 @@
-package transactions_show_test
-
-import (
-	"testing"
-)
-
-func TestIntegrationTransactionsShow(t *testing.T) {
-
-	//transactionsCommand, err := commands.New(cliRequest)
-	//testkit.AssertIsNil(t, err)
-	//
-	//err = transactionsCommand.Execute(map[string]interface{}{})
-	//testkit.AssertIsNil(t, err)
-	//
-	//How to i pass the repository and stuff... this is not the test i want.
-	//	do integration testing at the interactor level
-	//the commands should be tested by the acceptance tests... i think
-	//
-
-}
